Pedigree only:
* new option "u[integer]"
add a penalty weight (must use option y also) on genotyped individuals depending on the number of their genotyped children in order to penalize genotyping removals if the number of genotyped children is strictly greater than a given threshold.
* pedigree format extension
an individual with negative alleles means that its genotype is fixed to the absolute value of these alleles.