New options for BTD:
* B[integer] : (0) DFS, (1) BTD, (2) BTD-RDS, (3) RDS
RDS is equivalent to BTD-RDS except that it is based on path decomposition instead of tree decomposition.
* j[integer] : split larger clusters into a chain of smaller embedded clusters with a number of proper variables less than this number
Use this option (j1) in conjunction with B3 to get pure RDS search.
* E : merge leaf clusters with their fathers if they have a small local treewidth (to be used in conjunction with option e)
* r[integer] : limit on maximum cluster separator size. if a separator is greater then merge the two clusters.

BTD, BTD-RDS, and RDS reuse last best solution found as value choice heuristic. If option x is used, an initial solution can be passed to guide value choice heuristic.