[feature] option -t=[integer] max RPC in preprocessing with a given limit in memory space (in MBytes)
option -p=[integer] 
is now compatible with solution retrieval.

if using a negative number i (i<=-3) then it automatically selects an elimination bound lower equal to the median degree of the primal graph if this bound is lower than -i otherwise does nothing
moreover, a limit in memory space (in MBytes) is enforced.